services/web/requirements.txt
Flask
# Flask-SQLAlchemy
Flask-Migrate
Flask-WTF
gunicorn
psycopg2-binary
WTForms-SQLAlchemy
flask-jwt-extended
flask-restful
Flask-Script
marshmallow
